The Song class is now abstract and all derived classes must implement the abstract method "OpenStream", that returns a stream with the audio data.
The audio data type is determined by the "AudioType" property in the song class. Currently there is only a MP3 and WAV stream supported.

The audio player has been rewritten to support this new mode of filesystem-independent audio streaming,
so that later we can easily implement a audio stream playback from a youtube video.