When the user adds files to the library, a progress indicator now shows the current progress with a progress bar, the amount of songs added yet, and the total amount of songs.
The main changes were made the the LocalSongFinder class, because I needed two separate threads, one that searched for the songs on the file system and the other, to read the tags from the files.
The searcher thread adds the paths of the files with allowed extensions to a queue, while the other thread takes these paths from the queue, reads the tags from the file and adds the song to a list.

All this was implemented with save locking mechanisms, so that no collection has a cross-thread access.