Fixed a bug, that caused the YoutubeAudioPlayer to raise the SongFinished event, even if the song hasn't started yet.
This was caused by the VlcPlayer.EndReached event, that raised prematurely.
For now, the solution I've come up is a bit of a hack, i use the TimeChanged event of the VlcPlayer class and check if the current time is equal or greater than the total duration,
but because it doesn't fit always perfectly, I subtract one second from the total duration so that the SongFinished event is hopefully raised.