Completely fixed the playback process
There was a problem where a new song would be started as soon as an AudioPlayer was disposed.
This has been fixed by adding a new AudioPlayerState called "Stopped", which indicated that the playback has been stopped prematurely, while the "Finished" state indicates that a song has ended.
Now the next song is only played when the PlaybackState is set to "Finished"